refactoring most of the architecture
I've decided to abandon the pure-data approach to describing sites. It
turns out that it was mostly a lie I was telling myself anyhow; there
is inherently code involved in building the site definition.

The new approach is centered around polymorphic sources and paths, and
using function composition and mapping to build more complex behavior.