move protocols into their own namespace
I'm doing this to avoid problems with cider and code reloading. It's a
big disappointment that I have to reference the protocol functions
directly from the proto namespace. I'd prefer to reference them from
the respective saison.path and saison.source namespaces, but to
achieve that I'd have to figure out how to easily re-export vars with
their original metadata. This is good enough for now.